feat: minor fix for testing
