+++ packages/circuits/src/ReportIdentityProof.ts
@@ -1,12 +1,11 @@
-import { Prover } from '@unirep/circuits'
+import { BaseProof, Circuit, Prover } from '@unirep/circuits'
 import { Groth16Proof } from 'snarkjs'
-import { UnirepSocialCircuit } from '../test/types'
-import { UnirepSocialBaseProof } from './UnirepSocialBaseProof'
+import { UnirepSocialCircuit } from './types'
 
 /**
  * Verify the voter identity for voting the report
  */
-export class ReportIdentityProof extends UnirepSocialBaseProof {
+export class ReportIdentityProof extends BaseProof {
     readonly input = {
         reportNullifier: 0,
         attesterId: 1,
@@ -29,6 +28,6 @@ export class ReportIdentityProof extends UnirepSocialBaseProof {
         this.attesterId = this.publicSignals[this.input.attesterId]
         this.epoch = this.publicSignals[this.input.epoch]
         this.stateTreeRoot = this.publicSignals[this.input.stateTreeRoot]
-        this.unirepSocialCircuit = UnirepSocialCircuit.reportIdentityProof
+        this.circuit = UnirepSocialCircuit.reportIdentityProof as any as Circuit
     }
 }
