--- packages/circuits/src/UnirepSocialBaseProof.ts
@@ -1,6 +1,0 @@
-import { BaseProof } from '@unirep/circuits'
-import { UnirepSocialCircuit } from '../test/types'
-
-export class UnirepSocialBaseProof extends BaseProof {
-    protected unirepSocialCircuit?: UnirepSocialCircuit
-}
