+++ packages/circuits/src/types.ts
@@ -1,4 +1,4 @@
-import { Identity } from '@semaphore-protocol/identity'
+import { Identity } from '@semaphore-protocol/identity';
 
 export type IdentityObject = { hashUserId: string; id: Identity }
 
@@ -10,7 +10,7 @@ export type EpochKeyControl = {
     chainId: bigint
 }
 
-export enum UnirepSocialCircuit {
+export enum  UnirepSocialCircuit {
     reportIdentityProof = 'reportIdentityProof',
     reportNullifierProof = 'reportNullifierProof',
     reportNonNullifierProof = 'reportNonNullifierProof',
