+++ packages/circuits/test/utils.ts
@@ -1,18 +1,11 @@
 import { Identity } from '@semaphore-protocol/identity'
 import { Circuit, CircuitConfig } from '@unirep/circuits'
 import * as utils from '@unirep/utils'
-import {
-    ATTESTER_ID_BITS,
-    CHAIN_ID_BITS,
-    EPOCH_BITS,
-    NONCE_BITS,
-    REVEAL_NONCE_BITS,
-} from '@unirep/utils'
 import crypto from 'crypto'
 import { poseidon1, poseidon2 } from 'poseidon-lite'
 import { defaultProver } from '../provers/defaultProver'
+import { IdentityObject } from '../src/types'
 import { ProofGenerationError } from './error'
-import { EpochKeyControl, IdentityObject } from './types'
 
 export const genProofAndVerify = async (
     circuit: Circuit | string,
@@ -160,35 +153,6 @@ export const genReportIdentityCircuitInput = (config: {
         state_tree_root: stateTreeRoot,
     }
     return utils.stringifyBigInts(circuitInputs)
-}
-
-export const shiftBits = (
-    data: bigint,
-    shiftBits: bigint,
-    variableBits: bigint
-): bigint => {
-    return (data >> shiftBits) & ((BigInt(1) << variableBits) - BigInt(1))
-}
-
-export const decodeEpochKeyControl = (control: bigint): EpochKeyControl => {
-    let accBits = BigInt(0)
-    const nonce = shiftBits(control, accBits, NONCE_BITS)
-    accBits += NONCE_BITS
-    const epoch = shiftBits(control, accBits, EPOCH_BITS)
-    accBits += EPOCH_BITS
-    const attesterId = shiftBits(control, accBits, ATTESTER_ID_BITS)
-    accBits += ATTESTER_ID_BITS
-    const revealNonce = shiftBits(control, accBits, REVEAL_NONCE_BITS)
-    accBits += REVEAL_NONCE_BITS
-    const chainId = shiftBits(control, accBits, CHAIN_ID_BITS)
-
-    return {
-        nonce,
-        epoch,
-        attesterId,
-        revealNonce,
-        chainId,
-    }
 }
 
 export const randomData = () => [
